define gateway messaging pattern
