+++ ln-gateway/src/ln.rs
@@ -1,8 +1,6 @@
 use async_trait::async_trait;
 use fedimint_server::modules::ln::contracts::Preimage;
 
-use crate::GatewayRequestTrait;
-
 #[async_trait]
 pub trait LnRpc: Send + Sync + 'static {
     /// Attempt to pay an invoice and block till it succeeds, fails or times out
@@ -16,13 +14,3 @@ pub trait LnRpc: Send + Sync + 'static {
 
 #[derive(Debug)]
 pub struct LightningError(pub Option<i32>);
-
-#[async_trait]
-pub trait GatewayRpcSender: Send + Sync + 'static {
-    async fn send<R: GatewayRequestTrait>(&self, message: R) -> Result<R::Response, anyhow::Error>;
-}
-
-#[async_trait]
-pub trait GatewayRpcReceiver {
-    async fn receive(&mut self) -> ();
-}
